+++ js/index.js
@@ -44,3 +44,5 @@ export * from "./scoreMarkers.js";
 export * from "./labelCells.js";
 
 export * from "./MultiMatrix.js";
+
+export * from "./realizeFile.js";